do not rely on assembly internals
The assembly code was relying on the DX register
and given how hard it was to understand where the DX register comes from
(it's documented nowhere in Go Assembler materials), I believe it's more
robust to rely on parameter passing rather than function value internals
at the assembly level.

jitcall now requires its first parameter to be the address of the jitcode.
The rest of the parameters are passed in RDI as it was the case prior.

Signed-off-by: Tibor Vass <teabee89@gmail.com>